// qiankun-handleCode/src/enum.ts
export enum EAppStatus {
  NOT_LOADED = 'NOT_LOADED',
  LOADING = 'LOADING',
  LOADED = 'LOADED',
  BOOTSTRAPING = 'BOOTSTRAPING', // 初始化中
  NOT_MOUNTED = 'NOT_MOUNTED', // 挂载失败
  MOUNTING = 'MOUNTING',
  UNMOUNTED = 'UNMOUNTED',
  MOUNTED = 'MOUNTED',
  UNMOUNTING = 'UNMOUNTING',
}